+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.1] - 2020-XX-XX
 
+- Drawer menu structure
+- Splash screen with flare animation
+
 ### Added
 
 - CHANGELOG.md